[chore]폴더정리
